pyproject.toml
[project]
name = "pydantic-ai-mlx-lm"
version = "0.1.4"
description = "Adds MLX models support to Pydantic AI"
readme = "README.md"
authors = [{ name = "Doruk Gezici", email = "doruk@gezici.me" }]
requires-python = ">=3.12"
dependencies = ["mlx-lm>=0.21.1", "openai>=1.59.9", "pydantic-ai-slim>=0.0.19"]

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "inline-snapshot>=0.19.3",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venv = ".venv"
venvPath = "."
typeCheckingMode = "strict"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.inline-snapshot]
format-command = "ruff format --stdin-filename {filename}"
